api/requirements.txt
uvicorn
fastapi
sqlalchemy
pymysql
mysql-connector-python
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
azure-identity
azure-keyvault-secrets
azure-storage-blob
langchain-community
unstructured
langchain
astrapy>=0.5.3
langchain-community
langchain-together
langchain-astradb
langchainhub
fastembed